# ARM64: Optimize IndexOf(byte) and IndexOf(char) APIs using intrinsics.