@TamarChristinaArm 2020-06-09
> > @kunalspathak the answer is... it depends.. on older cores such as Cortex-A55 your sequence will probably be faster, on newer cores like Cortex-A76 (which is what AoR is tuned for) the `addp` version will beat the `uminv` one by a decent margin. `uminv` is a reasonably expensive operation and `addp` has been made cheaper. Even though it has more instructions the `addp` one will take less time to get through the pipeline.
> 
> Thanks @TamarChristinaArm. So probably, switching to the other version will be sensible to do?

I would say yes, even though you may get a lower increased based on hardware you may be benchmarking on now, on the long run it would the better sequence.