@AntonLapounov 2020-06-09
> `uminv` is a reasonably expensive operation and `addp` has been made cheaper.

@TamarChristinaArm Is there any table explaining instruction costs?