@kunalspathak 2020-06-09
> @TamarChristinaArm Is there any table explaining instruction costs?

So far I was referring to [Cortex-A55 optimization guide](https://static.docs.arm.com/epm128372/30/arm_cortex_a55_software_optimization_guide_v3.pdf), but looking at [Cortex-A76 optimization guide](https://static.docs.arm.com/swog307215/a/Arm_Cortex-A76_Software_Optimization_Guide.pdf), `uminv` takes 6 cycles vs. `addp` that takes 2.