@kunalspathak 2020-06-09
> Overall LGTM, just a few questions/comments 😄

Thanks @tannergooding . As Tamar suggested, I am going to switch over to the other implementation and then address your comments.