@kunalspathak 2020-06-11
> I do not think we need to use a mask at all. On input `vector` has 4 elements, each of them is either `0000` or `ffff`. Would the following work?
> 
> ```cs
> var zip = AdvSimd.Arm64.AddPairwise(vector.AsByte(), vector.AsByte()); // Sequence of 00 and fe bytes, two identical copies
> var scalar = zip.AsUInt64().ToScalar(); // Number like 0x00fe00fefe000000, each char corresponds to 1 byte = 8 bits
> return BitOperations.TrailingZeroCount(scalar) >> 3;
> ```

Thanks for your suggestion Anton. Did you mean to also `zip` something before given that you collect the result of `AddPairwise` inside `zip` variable. Regardless, we need 2 `AddPairwise` to aggregate the result  into 0th lane so we can use `AsUInt64().ToScalar()` to extract it out. The mask is needed to identify the lanes. If we don't use mask, the `AddPairwise` could add the consecutive set bits resulting in lost of information about the lane.