@AntonLapounov 2020-06-11
@kunalspathak Why do we need two `AddPairwise`? One `AddPairwise` reduces the size from 128 bit to 64 bit, which gives you the scalar.  Two `AddPairwise` would give you a 32 bit value.  And there is no information lost since my `AddPairwise` adds two _identical_ bytes.