@kunalspathak 2020-06-11
> @kunalspathak Why do we need two `AddPairwise`? One `AddPairwise` reduces the size from 128 bit to 64 bit, which gives you the scalar. Two `AddPairwise` would give you a 32 bit value. And there is no information lost since my `AddPairwise` adds two _identical_ bytes.

Ah, I see what you mean. I was thinking that your code is for `Byte` but it is for `char`. Yes, you are right, we don't need 2nd `AddPairwise`. However for `Byte` we need a mask to find which lane is selected, but I can get rid of  1 `AddPairwise`. Here is what it looks for `byte` now:

```csharp
var themask = Vector128.Create((byte)1, 16, 1, 16, 1, 16, 1, 16, 1, 16, 1, 16, 1, 16, 1, 16);
var masked = AdvSimd.And(vector, themask);
var zip = AdvSimd.Arm64.AddPairwise(masked, masked);
return BitOperations.TrailingZeroCount(zip.AsUInt64().ToScalar()) >> 2;
```