@AntonLapounov 2020-06-11
@kunalspathak Yes, I was thinking about the `char` case only.  Sorry for not being clear.