@kunalspathak 2020-06-12
> Looks great to me. One related thing I noticed is that the `Contains` method is not optimized the same way as the `IndexOf` method while a comment on the former says "Adapted from IndexOf(...)". I am wondering if we need to update `Contains` methods in these two files as well. @danmosemsft

`IndexOf` got optimized in https://github.com/dotnet/coreclr/pull/22118 but `Contains` was never optimized. While working on `IndexOfAny(byte)`, I noticed that we don't optimize `IndexOfAny(char)` too using x86 intrinsics.