@kunalspathak 2020-06-12
> > Looks great to me. One related thing I noticed is that the `Contains` method is not optimized the same way as the `IndexOf` method while a comment on the former says "Adapted from IndexOf(...)". I am wondering if we need to update `Contains` methods in these two files as well. @danmosemsft
> 
> `IndexOf` got optimized in [dotnet/coreclr#22118](https://github.com/dotnet/coreclr/pull/22118) but `Contains` was never optimized. While working on `IndexOfAny(byte)`, I noticed that we don't optimize `IndexOfAny(char)` too using x86 intrinsics.

Found following tracking issues:
- Use Intrinsics.X86 for SpanHelpers.SequenceCompareTo(char) dotnet/runtime#12093
- Use Intrinsics.X86 for SpanHelpers.IndexOfAny(char,char) dotnet/runtime#12094
- Use Intrinsics.X86 for SpanHelpers.IndexOfAny(char,char,char) dotnet/runtime#12095
- Use Intrinsics.X86 for SpanHelpers.IndexOfAny(char,char,char,char) dotnet/runtime#12096
- Use Intrinsics.X86 for SpanHelpers.IndexOfAny(char,char,char,char,char) dotnet/runtime#12097